refact: separando entidades de serviços
