+++ backend/domain/route/route.py
@@ -1,6 +1,4 @@
-from backend.domain.repositories.routes_repository import RoutesRepository
-
-class RoutesImpl:
+class Route:
 
     route_id = ''
     route_short_name = ''
@@ -36,11 +34,3 @@ class RoutesImpl:
             self.final_stop_id = ''
         else:
             self.final_stop_id = final_stop_id
-
-    def get_routes_from_stop(stop_id):
-        stop_routes_list = []
-        all_routes_from_stop = RoutesRepository.get_routes_from_stop_repo(stop_id)
-        for entry in all_routes_from_stop:
-            route = RoutesRepository.get_route_by_id_repo(entry.route_id)
-            stop_routes_list.append([entry.route_stop_id, RoutesImpl(entry.route_id, route.route_short_name)])
-        return stop_routes_list
